implemented graphical functions and modified window to close with its assigned task

+++ general.h
@@ -19,6 +19,6 @@
 #define clamp(x, l, r) min(max(x, l), r)
 
 typedef unsigned int uint;
-typedef uint size_t;
+typedef long unsigned int size_t;
 
 #endif
